+++ cli/common/bins.js
@@ -22,7 +22,7 @@ export const {
     ninja_exe,
     rescript_editor_analysis_exe,
     rescript_tools_exe,
+    rescript_legacy_exe,
     rescript_exe,
-    rewatch_exe,
   },
 } = mod;